+++ s3torchconnector/src/s3torchconnector/_s3client/s3client_config.py
@@ -20,6 +20,8 @@ class S3ClientConfig:
     force_path_style(bool): forceful path style addressing for S3 client.
     max_attempts(int): amount of retry attempts for retrieable errors.
     profile(str): Profile name to use for S3 authentication.
+    requester_pays(bool): Set to true to send the `x-amz-request-payer: requester` header on all
+        S3 requests, enabling access to Requester Pays buckets.
     """
 
     throughput_target_gbps: float = 10.0
@@ -28,3 +30,4 @@ class S3ClientConfig:
     force_path_style: bool = False
     max_attempts: int = 10
     profile: Optional[str] = None
+    requester_pays: bool = False