+++ s3torchconnectorclient/rust/src/mountpoint_s3_client.rs
@@ -52,6 +52,8 @@ pub struct MountpointS3Client {
     endpoint: Option<String>,
     #[pyo3(get)]
     max_attempts: usize,
+    #[pyo3(get)]
+    requester_pays: bool,
 }
 
 /// Waits for all managed CRT threads to complete, with a specified timeout.
@@ -99,7 +101,7 @@ pub fn join_all_managed_threads(py: Python<'_>, timeout_secs: f64) -> PyResult<(
 #[pymethods]
 impl MountpointS3Client {
     #[new]
-    #[pyo3(signature = (region, user_agent_prefix="".to_string(), throughput_target_gbps=10.0, part_size=8*1024*1024, profile=None, unsigned=false, endpoint=None, force_path_style=false, max_attempts=10))]
+    #[pyo3(signature = (region, user_agent_prefix="".to_string(), throughput_target_gbps=10.0, part_size=8*1024*1024, profile=None, unsigned=false, endpoint=None, force_path_style=false, max_attempts=10, requester_pays=false))]
     #[allow(clippy::too_many_arguments)]
     pub fn new_s3_client(
         region: String,
@@ -111,6 +113,7 @@ impl MountpointS3Client {
         endpoint: Option<String>,
         force_path_style: bool,
         max_attempts: usize,
+        requester_pays: bool,
     ) -> PyResult<Self> {
         // TODO: Mountpoint has logic for guessing based on instance type. It may be worth having
         // similar logic if we want to exceed 10Gbps reading for larger instances
@@ -135,13 +138,16 @@ impl MountpointS3Client {
             user_agent_string = &user_agent_prefix;
         }
 
-        let config = S3ClientConfig::new()
+        let mut config = S3ClientConfig::new()
             .user_agent(UserAgent::new(Some(user_agent_string.to_owned())))
             .throughput_target_gbps(throughput_target_gbps)
             .part_size(part_size)
             .auth_config(auth_config)
             .endpoint_config(endpoint_config)
             .max_attempts(NonZeroUsize::try_from(max_attempts).expect("max_attempts must be > 0"));
+        if requester_pays {
+            config = config.request_payer("requester");
+        }
         let crt_client = Arc::new(S3CrtClient::new(config).map_err(python_exception)?);
 
         Ok(MountpointS3Client::new(
@@ -155,6 +161,7 @@ impl MountpointS3Client {
             max_attempts,
             crt_client,
             endpoint,
+            requester_pays,
         ))
     }
 
@@ -238,6 +245,7 @@ impl MountpointS3Client {
             slf.endpoint.clone().into_pyobject(py)?.into_any(),
             slf.force_path_style.into_py_any(py)?.bind(py).to_owned(),
             slf.max_attempts.into_pyobject(py)?.into_any(),
+            slf.requester_pays.into_py_any(py)?.bind(py).to_owned(),
         ];
         PyTuple::new(py, state)
     }
@@ -257,6 +265,7 @@ impl MountpointS3Client {
         max_attempts: usize,
         client: Arc<Client>,
         endpoint: Option<String>,
+        requester_pays: bool,
     ) -> Self
     where
         Client: ObjectClient + Sync + Send + 'static,
@@ -274,6 +283,7 @@ impl MountpointS3Client {
             client: Arc::new(MountpointS3ClientInnerImpl::new(client)),
             user_agent_prefix,
             endpoint,
+            requester_pays,
         }
     }
 }